third_party/blink/web_tests/virtual/third-party-storage-partitioning/external/wpt/workers/shared-worker-partitioned.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] Test partitioning of shared workers
  assert_equals: expected 1 but got 2
Harness: the test ran to completion.
